+++ CHANGELOG.md
@@ -69,6 +69,9 @@ project.
   (https://mad-srv.informatik.uni-erlangen.de/MadLab/GaitAnalysis/gaitmap/-/merge_requests/138)
 
 ### Fixed
+- After internal updates of scipy>=1.6 we now manually handle empty orientation inputs in 
+`gaitmap.parameters.spatial_parameters._calc_turning_angle`.
+(https://mad-srv.informatik.uni-erlangen.de/MadLab/GaitAnalysis/gaitmap/-/merge_requests/153)
 
 ### Migration Guide